Move benchmark lifecycle projection inward
